MIL. Debugged the structure.
Debugged those parts of the code that do not directly pertain to the
algorithmic part, such as geometry of the network, and "appearances"
(presence of NaN values in the output, etc.).

The validity of the implementation of the Back Propagation Algorithm
itself is yet to be established through practical application.